chore: tidy up build processes
