+++ package.json
@@ -9,40 +9,48 @@
     "start": "webpack --watch --progress --mode=development",
     "start:profiles": "webpack-dev-server --progress --mode=development --config webpack.profiles.config.js --inline",
     "start:firefox": "webpack --watch --progress --mode=development --env.PLATFORM=firefox",
-    "build:staging": "NODE_ENV=staging yarn build:chromium && NODE_ENV=staging yarn build:firefox",
-    "release:staging": "NODE_ENV=staging yarn build:chromium --env.SENTRY_ENABLED && NODE_ENV=staging yarn build:firefox --env.SENTRY_ENABLED",
-    "build:proding": "NODE_ENV=proding yarn build:chromium && NODE_ENV=proding yarn build:firefox",
-    "release:proding": "NODE_ENV=proding yarn build:chromium --env.SENTRY_ENABLED && NODE_ENV=proding yarn build:firefox --env.SENTRY_ENABLED",
+    "build:chromium": "webpack --mode=production --env.PLATFORM=chromium",
     "build:firefox": "webpack --mode=production --env.PLATFORM=firefox",
-    "sign:firefox": "node --experimental-modules bin/web-ext-sign.mjs",
-    "sign:firefox:staging": "NODE_ENV=staging yarn sign:firefox",
-    "sign:firefox:proding": "NODE_ENV=proding yarn sign:firefox",
-    "sign:firefox:production": "NODE_ENV=production yarn sign:firefox",
-    "upload:chrome": "node --experimental-modules bin/chrome-webstore-upload",
+    "build:chromium:staging": "NODE_ENV=staging yarn build:chromium",
+    "build:chromium:proding": "NODE_ENV=proding yarn build:chromium",
+    "build:chromium:production": "NODE_ENV=production yarn build:chromium",
+    "build:firefox:staging": "NODE_ENV=staging yarn build:firefox",
+    "build:firefox:proding": "NODE_ENV=proding yarn build:firefox",
+    "build:firefox:production": "NODE_ENV=production yarn build:firefox",
+    "build:staging": "yarn build:chromium:staging && yarn build:firefox:staging",
+    "build:proding": "yarn build:chromium:proding && yarn build:firefox:proding",
+    "build:production": "yarn build:chromium:production && yarn build:firefox:production",
+    "buildVersion:chromium:staging": "yarn build:chromium:staging --env.SENTRY_ENABLED",
+    "buildVersion:chromium:proding": "yarn build:chromium:proding --env.SENTRY_ENABLED",
+    "buildVersion:chromium:production": "yarn build:chromium:production --env.SENTRY_ENABLED",
+    "buildVersion:firefox:staging": "yarn build:firefox:staging --env.SENTRY_ENABLED",
+    "buildVersion:firefox:proding": "yarn build:firefox:proding --env.SENTRY_ENABLED",
+    "buildVersion:firefox:production": "yarn build:firefox:production --env.SENTRY_ENABLED",
+    "buildVersion:staging": "yarn buildVersion:chromium:staging && yarn buildVersion:firefox:staging",
+    "buildVersion:proding": "yarn buildVersion:chromium:proding && yarn buildVersion:firefox:proding",
+    "buildVersion:production": "yarn buildVersion:chromium:production && yarn buildVersion:firefox:production",
+    "buildVersion": "yarn release:staging && yarn release:proding && yarn release:production",
+    "upload:firefox:_": "node --experimental-modules bin/web-ext-sign.mjs",
+    "upload:firefox:staging": "NODE_ENV=staging yarn upload:firefox:_",
+    "upload:firefox:proding": "NODE_ENV=proding yarn upload:firefox:_",
+    "upload:firefox:production": "NODE_ENV=production yarn upload:firefox:_",
     "upload:chrome:staging": "NODE_ENV=staging node --experimental-modules bin/chrome-webstore-upload.mjs",
     "upload:chrome:proding": "NODE_ENV=proding node --experimental-modules bin/chrome-webstore-upload.mjs",
     "upload:chrome:production": "NODE_ENV=production node --experimental-modules bin/chrome-webstore-upload.mjs",
-    "build:chromium": "webpack --mode=production --env.PLATFORM=chromium",
-    "build:production": "NODE_ENV=production yarn build:chromium && NODE_ENV=production yarn build:firefox",
-    "release:production": "NODE_ENV=production yarn build:chromium --env.SENTRY_ENABLED && NODE_ENV=production yarn build:firefox --env.SENTRY_ENABLED",
-    "release:production:firefox": "NODE_ENV=production yarn build:firefox --env.SENTRY_ENABLED",
+    "release": "yarn semantic-release",
     "build": "yarn build:staging && yarn build:production && yarn build:proding",
-    "release": "yarn release:staging && yarn release:production && yarn release:proding",
     "clean": "rm -rf build",
     "test": "BACKEND_ORIGIN='' mocha -r ts-node/register -r tsconfig-paths/register -r jsdom-global/register 'src/**/*.spec.ts' 'test/app/*.ts' 'test/app/**/*.ts'",
     "typecheck": "tsc --noEmit",
     "lint": "yarn lint:ts",
-    "prettier": "prettier --write src/**/*.{ts,tsx} test/**/*.{ts,tsx,js} manifest/**/*.js webpack/*.js release.*.js",
     "lint-fixme-stylelint-has-to-be-fixed": "yarn lint:ts && yarn lint:css",
     "lint:ts": "eslint 'src/**/*.{ts,tsx}'",
     "lint:css": "stylelint './src/**/*.{ts,tsx}'",
+    "prettier": "prettier --write src/**/*.{ts,tsx} test/**/*.{ts,tsx,js} manifest/**/*.js webpack/*.js release.*.js",
     "verify": "yarn test && yarn typecheck && yarn lint && yarn build:production && yarn build-storybook",
     "analyze": "NODE_ENV=production webpack --mode=production --env.PLATFORM=chromium --env.ANALYZE",
     "storybook": "start-storybook --ci -p 6007 -c .storybook",
-    "build-storybook": "build-storybook",
-    "deploy": "DEPLOY=staging yarn semantic-release",
-    "deploy:production": "DEPLOY=production yarn semantic-release",
-    "deploy:chrome": "yarn run upload:chrome:staging && yarn run upload:chrome:proding"
+    "build-storybook": "build-storybook"
   },
   "husky": {
     "hooks": {
@@ -59,10 +67,10 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/insitu-project/recommendations-webextension"
+    "url": "https://github.com/dis-moi/extension"
   },
-  "homepage": "https://www.bulles.fr",
-  "author": "INSITU SAS (https://github.com/insitu-project)",
+  "homepage": "https://www.dismoi.io",
+  "author": "DisMoi (https://github.com/dis-moi)",
   "contributors": [
     "Benjamin Menant (https://github.com/bmenant)",
     "Arnaud Lemaire (https://github.com/lilobase)",
@@ -70,7 +78,10 @@
     "Romain l'Ourson (https://github.com/iOiurson)",
     "Vincent Van Steenbergen (https://github.com/nsteenv)",
     "Mihail Diordiev <zalmoxisus@gmail.com> (https://github.com/zalmoxisus)",
-    "Jhen <developer@jhen.me> (https://github.com/jhen0409)"
+    "Jhen <developer@jhen.me> (https://github.com/jhen0409)",
+    "Erick Gardin <erick@entre-quote.com> (https://github.com/newick)",
+    "Johan Dufour <johan.dufour@gmail.com> (https://github.com/lutangar)",
+    "Jalil Arfaoui<jalil@arfaoui.net> (https://github.com/jalilarfaoui)"
   ],
   "license": "GPL-3.0",
   "dependencies": {
